Fixed a bug with setWaterLevel method in SystemData. Should take int values instead of double. Aquaponics fully functional now in conjunction with a fix in the back-end system.

+++ systemh-android/src/edu/hawaii/systemh/android/systemdata/SystemData.java
@@ -383,7 +383,7 @@ public class SystemData {
    * 
    * @param value Desired water level
    */
-  public void setWaterLevel(double value) {
+  public void setWaterLevel(int value) {
 
     try {
       String putUrl =
